refactor: modulariza conteúdos das trilhas e melhora renderização das apostilas
