+++ Front Lab Academy/src/data/roteiro.js
@@ -1,1 +1,110 @@
-export const roadmap = []
+export const roadmap = [
+  { title: '01. Como a web funciona (request/response)', trail: 'preparacao-do-aluno', mod: 0 },
+  { title: '02. Estrutura de projeto front-end', trail: 'preparacao-do-aluno', mod: 0 },
+  { title: '03. Setup do VS Code para produtividade', trail: 'preparacao-do-aluno', mod: 1 },
+  { title: '04. Extensões e snippets essenciais', trail: 'preparacao-do-aluno', mod: 1 },
+  { title: '05. Terminal: navegação e comandos base', trail: 'preparacao-do-aluno', mod: 2 },
+  { title: '06. Terminal: automação com scripts', trail: 'preparacao-do-aluno', mod: 2 },
+  { title: '07. Git: init, add, commit', trail: 'preparacao-do-aluno', mod: 3 },
+  { title: '08. GitHub: push, PR e revisão', trail: 'preparacao-do-aluno', mod: 3 },
+
+  { title: '09. HTML: estrutura base', trail: 'html', mod: 2 },
+  { title: '10. HTML: títulos e hierarquia', trail: 'html', mod: 5 },
+  { title: '11. Navegação: links externos e internos', trail: 'html', mod: 9 },
+  { title: '12. Navegação: âncoras na página', trail: 'html', mod: 10 },
+  { title: '13. Listas e conteúdo estruturado', trail: 'html', mod: 16 },
+  { title: '14. Imagens e texto alternativo', trail: 'html', mod: 12 },
+  { title: '15. Tabelas com semântica correta', trail: 'html', mod: 18 },
+  { title: '16. Formulários: campos e labels', trail: 'html', mod: 26 },
+  { title: '17. Formulários: agrupamento e validação', trail: 'html', mod: 28 },
+  { title: '18. Acessibilidade base em HTML', trail: 'html', mod: 40 },
+
+  { title: '19. CSS: conexão com HTML', trail: 'css', mod: 1 },
+  { title: '20. CSS: seletores e especificidade inicial', trail: 'css', mod: 3 },
+  { title: '21. Cores, contraste e unidade visual', trail: 'css', mod: 4 },
+  { title: '22. Tipografia e hierarquia', trail: 'css', mod: 7 },
+  { title: '23. Box model e espaçamento', trail: 'css', mod: 10 },
+  { title: '24. Width, height e overflow', trail: 'css', mod: 17 },
+  { title: '25. Flexbox: alinhamento e distribuição', trail: 'css', mod: 21 },
+  { title: '26. Flexbox: navbar e cards', trail: 'css', mod: 28 },
+  { title: '27. Grid: colunas, linhas e gaps', trail: 'css', mod: 31 },
+  { title: '28. Grid: composição de página', trail: 'css', mod: 37 },
+  { title: '29. Componentes visuais com CSS', trail: 'css', mod: 46 },
+  { title: '30. Responsividade mobile-first', trail: 'css', mod: 52 },
+  { title: '31. Projeto final com CSS responsivo', trail: 'css', mod: 59 },
+
+  { title: '32. JS: variáveis e tipos', trail: 'javascript-e-logica', mod: 0 },
+  { title: '33. JS: condicionais', trail: 'javascript-e-logica', mod: 0 },
+  { title: '34. JS: loops e iteração', trail: 'javascript-e-logica', mod: 1 },
+  { title: '35. JS: funções e escopo', trail: 'javascript-e-logica', mod: 1 },
+  { title: '36. Arrays: map/filter/reduce', trail: 'javascript-e-logica', mod: 2 },
+  { title: '37. Arrays: buscas e ordenação', trail: 'javascript-e-logica', mod: 2 },
+  { title: '38. Objetos: modelagem de dados', trail: 'javascript-e-logica', mod: 3 },
+  { title: '39. Objetos: destructuring e spread', trail: 'javascript-e-logica', mod: 3 },
+  { title: '40. DOM: seleção e atualização', trail: 'javascript-e-logica', mod: 4 },
+  { title: '41. Eventos: clique, input, submit', trail: 'javascript-e-logica', mod: 4 },
+  { title: '42. Fetch API: loading e erro', trail: 'javascript-e-logica', mod: 5 },
+  { title: '43. Fetch API: transformação de dados', trail: 'javascript-e-logica', mod: 5 },
+  { title: '44. LocalStorage: persistência de estado', trail: 'javascript-e-logica', mod: 6 },
+  { title: '45. LocalStorage: sincronização de UI', trail: 'javascript-e-logica', mod: 6 },
+
+  { title: '46. TypeScript: inferência e tipos primitivos', trail: 'typescript', mod: 0 },
+  { title: '47. TypeScript: strict mode', trail: 'typescript', mod: 0 },
+  { title: '48. Interfaces e type aliases', trail: 'typescript', mod: 1 },
+  { title: '49. Modelagem de domínio com interfaces', trail: 'typescript', mod: 1 },
+  { title: '50. Union types e narrowing', trail: 'typescript', mod: 2 },
+  { title: '51. Guards de tipo reutilizáveis', trail: 'typescript', mod: 2 },
+  { title: '52. Generics: funções e utilitários', trail: 'typescript', mod: 3 },
+  { title: '53. Generics com constraints', trail: 'typescript', mod: 3 },
+  { title: '54. Tipagem de eventos de UI', trail: 'typescript', mod: 4 },
+  { title: '55. Tipagem de callbacks e handlers', trail: 'typescript', mod: 4 },
+  { title: '56. DTOs e contratos de API', trail: 'typescript', mod: 5 },
+  { title: '57. Estrutura de tipos por domínio', trail: 'typescript', mod: 6 },
+
+  { title: '58. Frameworks: arquitetura por componentes', trail: 'frameworks-front-end', mod: 0 },
+  { title: '59. Frameworks: composição e reutilização', trail: 'frameworks-front-end', mod: 0 },
+  { title: '60. Roteamento: páginas e parâmetros', trail: 'frameworks-front-end', mod: 1 },
+  { title: '61. Roteamento: layouts e navegação', trail: 'frameworks-front-end', mod: 1 },
+  { title: '62. Estado compartilhado entre telas', trail: 'frameworks-front-end', mod: 2 },
+  { title: '63. Organização de estado por domínio', trail: 'frameworks-front-end', mod: 2 },
+  { title: '64. Formulários em framework', trail: 'frameworks-front-end', mod: 3 },
+  { title: '65. Validação e feedback de formulário', trail: 'frameworks-front-end', mod: 3 },
+  { title: '66. Consumo de API em SPA', trail: 'frameworks-front-end', mod: 4 },
+  { title: '67. Estados de loading/erro/vazio', trail: 'frameworks-front-end', mod: 4 },
+  { title: '68. Renderização e performance', trail: 'frameworks-front-end', mod: 5 },
+  { title: '69. SSR/SSG/CSR na prática', trail: 'frameworks-front-end', mod: 5 },
+  { title: '70. Arquitetura escalável de app', trail: 'frameworks-front-end', mod: 6 },
+
+  { title: '71. Git avançado para equipes', trail: 'ferramentas', mod: 0 },
+  { title: '72. Code review com Pull Request', trail: 'ferramentas', mod: 1 },
+  { title: '73. npm scripts para produtividade', trail: 'ferramentas', mod: 2 },
+  { title: '74. npm workspaces e organização', trail: 'ferramentas', mod: 2 },
+  { title: '75. Vite: aliases e envs', trail: 'ferramentas', mod: 3 },
+  { title: '76. Vite: build e preview', trail: 'ferramentas', mod: 3 },
+  { title: '77. ESLint: regras e consistência', trail: 'ferramentas', mod: 4 },
+  { title: '78. Prettier: formatação automática', trail: 'ferramentas', mod: 5 },
+  { title: '79. Pipeline local de qualidade', trail: 'ferramentas', mod: 6 },
+  { title: '80. Pipeline de release', trail: 'ferramentas', mod: 6 },
+
+  { title: '81. Acessibilidade: teclado e foco', trail: 'qualidade-e-boas-praticas', mod: 0 },
+  { title: '82. Acessibilidade: semântica avançada', trail: 'qualidade-e-boas-praticas', mod: 0 },
+  { title: '83. Performance: imagens e fontes', trail: 'qualidade-e-boas-praticas', mod: 1 },
+  { title: '84. Performance: render e bundle', trail: 'qualidade-e-boas-praticas', mod: 1 },
+  { title: '85. Responsividade robusta', trail: 'qualidade-e-boas-praticas', mod: 2 },
+  { title: '86. Testes unitários de regra', trail: 'qualidade-e-boas-praticas', mod: 3 },
+  { title: '87. Testes de interface e fluxo', trail: 'qualidade-e-boas-praticas', mod: 4 },
+  { title: '88. PWA: manifest e cache', trail: 'qualidade-e-boas-praticas', mod: 5 },
+  { title: '89. Governança e manutenção', trail: 'qualidade-e-boas-praticas', mod: 6 },
+
+  { title: '90. Projeto final: planejamento', trail: 'qualidade-e-boas-praticas', mod: 6 },
+  { title: '91. Projeto final: arquitetura', trail: 'frameworks-front-end', mod: 6 },
+  { title: '92. Projeto final: implementação UI', trail: 'css', mod: 59 },
+  { title: '93. Projeto final: integração de dados', trail: 'javascript-e-logica', mod: 5 },
+  { title: '94. Projeto final: tipagem', trail: 'typescript', mod: 5 },
+  { title: '95. Projeto final: validação e testes', trail: 'qualidade-e-boas-praticas', mod: 4 },
+  { title: '96. Projeto final: otimização de performance', trail: 'qualidade-e-boas-praticas', mod: 1 },
+  { title: '97. Projeto final: revisão de acessibilidade', trail: 'qualidade-e-boas-praticas', mod: 0 },
+  { title: '98. Projeto final: documentação técnica', trail: 'ferramentas', mod: 6 },
+  { title: '99. Projeto final: deploy e revisão final', trail: 'preparacao-do-aluno', mod: 3 },
+  { title: '100. Pronto para portfólio e entrevista', trail: 'qualidade-e-boas-praticas', mod: 6 }
+]
