+++ Front Lab Academy/src/data/trilhas/bootstrap/modulos.js
@@ -1,1 +1,12 @@
-export const bootstrapModules = []
+import { mkStarter } from '../../helpers/criar-starter.js'
+
+export const bootstrapModules = [
+      { title: 'Introdução ao Bootstrap', learn: ['Quando usar Bootstrap', 'Instalação via CDN/npm', 'Estrutura base do CSS e JS'], practice: 'Crie uma página com Bootstrap via CDN e valide se CSS e JS carregam corretamente.', exercise: 'Conclua quando navbar, main e footer usarem classes Bootstrap sem CSS customizado obrigatório.', starter: mkStarter('Bootstrap inicial', '#7952b3', 'Bootstrap pronto') },
+      { title: 'Containers, grid e breakpoints', learn: ['container e container-fluid', 'row e col', 'breakpoints responsivos'], practice: 'Construa uma seção que tenha uma coluna no mobile, duas no tablet e três no desktop.', exercise: 'Conclua quando o grid usar row, col e breakpoints sem largura fixa.', starter: mkStarter('Grid Bootstrap', '#6f42c1', 'Grid Bootstrap') },
+      { title: 'Utilitários de espaçamento e tipografia', learn: ['classes de margin/padding', 'display e flex utilities', 'tipografia e cores'], practice: 'Monte um card de perfil usando utilitários de padding, margin, cor e texto.', exercise: 'Conclua quando o componente ficar legível sem criar classes CSS novas.', starter: mkStarter('Utilitários Bootstrap', '#7952b3', 'Utilities') },
+      { title: 'Componentes essenciais', learn: ['navbar', 'cards', 'buttons', 'alerts', 'badges'], practice: 'Combine navbar, cards, badges e botões em uma tela de catálogo de cursos.', exercise: 'Conclua quando cada componente tiver função clara e texto coerente.', starter: mkStarter('Componentes Bootstrap', '#6f42c1', 'Components') },
+      { title: 'Formulários e validação visual', learn: ['form-control', 'input groups', 'feedback de validação'], practice: 'Crie um cadastro com input group, required e mensagens de feedback visual.', exercise: 'Conclua quando campos válidos e inválidos tiverem estados reconhecíveis.', starter: mkStarter('Forms Bootstrap', '#7952b3', 'Forms') },
+      { title: 'Componentes com JavaScript', learn: ['modal', 'dropdown', 'collapse', 'tabs'], practice: 'Implemente uma FAQ com collapse e um modal de confirmação para ação importante.', exercise: 'Conclua quando as interações funcionarem pelo Bootstrap sem script manual extra.', starter: mkStarter('Bootstrap JS', '#6f42c1', 'Bootstrap JS') },
+      { title: 'Customização e tema', learn: ['variáveis CSS', 'Sass', 'sobrescrita segura', 'design system simples'], practice: 'Altere cores, radius e botões mantendo o comportamento original dos componentes.', exercise: 'Conclua quando o tema parecer de uma marca própria sem quebrar contraste.', starter: mkStarter('Tema Bootstrap', '#7952b3', 'Theme') },
+      { title: 'Projeto final com Bootstrap', learn: ['composição de layout', 'responsividade completa', 'documentação do projeto'], practice: 'Planeje e construa uma landing ou painel usando grid, cards, formulário e modal.', exercise: 'Conclua quando houver deploy, README e screenshots mobile e desktop.', starter: mkStarter('Projeto Bootstrap', '#6f42c1', 'Projeto Bootstrap') }
+    ]
