+++ Front Lab Academy/src/data/trilhas/css/atividades.js
@@ -1,1 +1,242 @@
-export const cssModuleActivities = {}
+export const cssModuleActivities = {
+  'Introdução ao CSS': {
+    practice: 'Abra um HTML simples e marque quais partes serão estrutura e quais serão aparência.',
+    exercise: 'Conclua quando conseguir explicar, em três frases, o papel de HTML, CSS e JavaScript na mesma página.'
+  },
+  'Como conectar CSS no HTML': {
+    practice: 'Crie um arquivo style.css, conecte com link no head e altere cor de fundo, cor do texto e fonte do body.',
+    exercise: 'Conclua quando o CSS externo carregar sem usar style inline e sem erro no caminho do arquivo.'
+  },
+  'Seletores básicos': {
+    practice: 'Estilize h1, p e button usando apenas seletores de tag em uma página de apresentação.',
+    exercise: 'Conclua quando cada tipo de elemento tiver um estilo visível e nenhuma regra afetar elementos fora do esperado.'
+  },
+  'Seletores por classe, ID e tag': {
+    practice: 'Crie dois cards com a mesma classe e destaque apenas um deles usando um ID.',
+    exercise: 'Conclua quando souber justificar por que a classe foi usada para reaproveitar estilo e o ID para um caso único.'
+  },
+  'Cores no CSS': {
+    practice: 'Monte uma seção com título, parágrafo e botão usando color, background-color e uma cor de destaque.',
+    exercise: 'Conclua quando texto e fundo tiverem contraste confortável e a paleta usar no máximo três cores principais.'
+  },
+  'Unidades de medida': {
+    practice: 'Compare px, rem, % e vw ajustando largura de card, tamanho de fonte e espaçamento.',
+    exercise: 'Conclua quando a página continuar legível ao mudar o tamanho da janela.'
+  },
+  'Fontes e tipografia': {
+    practice: 'Defina uma fonte para o body e outra hierarquia visual para h1, h2 e parágrafos.',
+    exercise: 'Conclua quando a página tiver leitura consistente sem depender de muitas famílias diferentes.'
+  },
+  'Tamanho, peso e estilo de texto': {
+    practice: 'Crie uma escala simples com título, subtítulo, texto comum e texto auxiliar.',
+    exercise: 'Conclua quando a hierarquia for percebida por tamanho, peso e line-height, não só por cor.'
+  },
+  'Alinhamento de texto': {
+    practice: 'Monte três blocos curtos comparando texto à esquerda, centralizado e alinhado à direita.',
+    exercise: 'Conclua quando escolher o alinhamento mais legível para um card de conteúdo.'
+  },
+  'Espaçamento com margin e padding': {
+    practice: 'Crie dois cards e ajuste o espaço interno com padding e a distância entre eles com margin.',
+    exercise: 'Conclua quando conseguir apontar qual espaçamento está dentro do card e qual está fora.'
+  },
+  'O que é o Box Model': {
+    practice: 'Use DevTools para observar content, padding, border e margin em um card.',
+    exercise: 'Conclua quando o card tiver cada camada do box model identificável por estilo ou anotação.'
+  },
+  'Width e height': {
+    practice: 'Defina largura máxima para um card e altura mínima para uma seção sem travar o conteúdo.',
+    exercise: 'Conclua quando textos maiores não vazarem nem ficarem cortados.'
+  },
+  Border: {
+    practice: 'Adicione bordas diferentes em card, input e botão para separar estados visuais.',
+    exercise: 'Conclua quando as bordas ajudarem na leitura sem deixar a interface pesada.'
+  },
+  'Border-radius': {
+    practice: 'Aplique radius em card, botão e imagem, mantendo uma escala coerente entre eles.',
+    exercise: 'Conclua quando nenhum canto parecer aleatório ou diferente sem motivo.'
+  },
+  'Margin na prática': {
+    practice: 'Organize uma seção vertical com título, texto e dois cards usando margin para separar blocos.',
+    exercise: 'Conclua quando o ritmo vertical estiver regular e sem espaçamentos duplicados.'
+  },
+  'Padding na prática': {
+    practice: 'Aumente a área interna de botões e cards até ficarem confortáveis para leitura e clique.',
+    exercise: 'Conclua quando o conteúdo não encostar nas bordas em nenhum componente.'
+  },
+  'Box-sizing': {
+    practice: 'Compare um card com content-box e outro com border-box usando a mesma largura, padding e border.',
+    exercise: 'Conclua quando aplicar box-sizing: border-box no reset e explicar a diferença visual.'
+  },
+  Overflow: {
+    practice: 'Crie um card com texto longo e teste overflow hidden, auto e visible.',
+    exercise: 'Conclua quando escolher uma solução que preserve o conteúdo sem quebrar o layout.'
+  },
+  'Display block, inline e inline-block': {
+    practice: 'Compare span, a, p e button mudando o display e observando fluxo e tamanho.',
+    exercise: 'Conclua quando conseguir alinhar links como botões sem transformar a página em um layout rígido.'
+  },
+  'Reset CSS básico': {
+    practice: 'Crie um reset com box-sizing, margin do body, imagens responsivas e fonte base.',
+    exercise: 'Conclua quando uma página nova começar com espaçamentos previsíveis entre navegadores.'
+  },
+  'Introdução a layouts': {
+    practice: 'Desenhe a estrutura de uma página com header, conteúdo principal, lista de cards e footer.',
+    exercise: 'Conclua quando cada bloco tiver uma função clara antes de aplicar Flexbox ou Grid.'
+  },
+  'Display flex': {
+    practice: 'Transforme uma lista horizontal de três itens em um container flex.',
+    exercise: 'Conclua quando os itens ficarem alinhados em linha sem usar float ou position.'
+  },
+  'Flex-direction': {
+    practice: 'Altere uma seção de cards entre row e column para entender o eixo principal.',
+    exercise: 'Conclua quando souber montar a versão empilhada e a versão lado a lado do mesmo conteúdo.'
+  },
+  'Justify-content': {
+    practice: 'Distribua logo, links e botão dentro de uma barra usando justify-content.',
+    exercise: 'Conclua quando o espaço horizontal ficar equilibrado sem margins manuais em cada item.'
+  },
+  'Align-items': {
+    practice: 'Alinhe ícone, texto e botão dentro de um card horizontal.',
+    exercise: 'Conclua quando todos os elementos ficarem centralizados no eixo cruzado.'
+  },
+  Gap: {
+    practice: 'Substitua margens entre itens por gap em uma lista flexível.',
+    exercise: 'Conclua quando o espaçamento continuar correto mesmo ao adicionar ou remover itens.'
+  },
+  'Flex-wrap': {
+    practice: 'Crie uma fileira de tags que quebra linha quando falta espaço.',
+    exercise: 'Conclua quando nenhuma tag gerar rolagem horizontal em tela pequena.'
+  },
+  'Cards com Flexbox': {
+    practice: 'Monte três cards com título, texto e botão mantendo o botão no fim do card.',
+    exercise: 'Conclua quando cards com textos diferentes continuarem alinhados visualmente.'
+  },
+  'Navbar com Flexbox': {
+    practice: 'Crie uma navbar com logo, links e botão de ação usando apenas Flexbox.',
+    exercise: 'Conclua quando a navegação ficar alinhada e com espaçamento consistente.'
+  },
+  'Centralização com Flexbox': {
+    practice: 'Centralize um estado vazio com título, texto e botão no centro de uma seção.',
+    exercise: 'Conclua quando a centralização funcionar nos dois eixos sem position absolute.'
+  },
+  'Introdução ao CSS Grid': {
+    practice: 'Monte uma grade simples de quatro cards usando display grid.',
+    exercise: 'Conclua quando entender quais linhas e colunas o navegador criou.'
+  },
+  'Grid-template-columns': {
+    practice: 'Crie uma grade com três colunas usando fr, repeat e minmax em versões separadas.',
+    exercise: 'Conclua quando escolher a solução que melhor se adapta ao conteúdo.'
+  },
+  'Grid-template-rows': {
+    practice: 'Monte um layout com cabeçalho, conteúdo e rodapé controlando as linhas do grid.',
+    exercise: 'Conclua quando a área principal ocupar o espaço restante sem altura fixa desnecessária.'
+  },
+  'Gap no Grid': {
+    practice: 'Ajuste espaços horizontais e verticais em uma galeria usando gap, row-gap e column-gap.',
+    exercise: 'Conclua quando a grade tiver ritmo regular sem margin nos cards.'
+  },
+  'Grid com cards': {
+    practice: 'Crie uma vitrine de cards com repeat e colunas proporcionais.',
+    exercise: 'Conclua quando os cards ficarem alinhados e fáceis de comparar.'
+  },
+  'Grid responsivo básico': {
+    practice: 'Use repeat(auto-fit, minmax()) para criar cards que se reorganizam sozinhos.',
+    exercise: 'Conclua quando a grade funcionar em celular e desktop sem media query.'
+  },
+  'Grid-area': {
+    practice: 'Nomeie áreas de header, sidebar, main e footer em um layout de documentação.',
+    exercise: 'Conclua quando o CSS revelar claramente onde cada área fica na página.'
+  },
+  'Layout de página com Grid': {
+    practice: 'Monte uma página com sidebar e conteúdo principal usando grid-template-areas.',
+    exercise: 'Conclua quando o layout tiver hierarquia clara e não depender de posicionamento manual.'
+  },
+  'Comparação entre Flexbox e Grid': {
+    practice: 'Resolva o mesmo bloco de cards com Flexbox e com Grid, anotando a diferença.',
+    exercise: 'Conclua quando souber dizer qual solução ficou mais simples para aquele caso.'
+  },
+  'Quando usar Flexbox ou Grid': {
+    practice: 'Escolha Flexbox para um componente e Grid para uma seção completa da mesma página.',
+    exercise: 'Conclua quando cada escolha tiver uma justificativa ligada ao eixo do layout.'
+  },
+  Backgrounds: {
+    practice: 'Crie uma seção com fundo sólido e outra com imagem de fundo posicionada.',
+    exercise: 'Conclua quando o texto continuar legível sobre qualquer fundo usado.'
+  },
+  'Imagens no CSS': {
+    practice: 'Use background-image em um banner decorativo sem colocar conteúdo importante na imagem.',
+    exercise: 'Conclua quando o HTML continuar compreensível mesmo sem carregar a imagem.'
+  },
+  'Object-fit': {
+    practice: 'Monte três cards com imagens de tamanhos diferentes usando object-fit: cover.',
+    exercise: 'Conclua quando todas as imagens preencherem o mesmo espaço sem distorção.'
+  },
+  'Sombras com box-shadow': {
+    practice: 'Aplique sombras leves em cards e compare estados normal e hover.',
+    exercise: 'Conclua quando a sombra indicar profundidade sem reduzir contraste.'
+  },
+  'Text-shadow': {
+    practice: 'Aplique text-shadow em um título sobre imagem e teste a leitura.',
+    exercise: 'Conclua quando a sombra melhorar contraste sem parecer decoração exagerada.'
+  },
+  'Botões estilizados': {
+    practice: 'Crie botão primário, secundário e desabilitado com tamanhos e estados coerentes.',
+    exercise: 'Conclua quando cada botão comunicar sua prioridade sem depender só de texto.'
+  },
+  'Cards modernos': {
+    practice: 'Monte um card de curso com tag, título, descrição, tempo e botão.',
+    exercise: 'Conclua quando o card tiver hierarquia clara e espaçamento consistente.'
+  },
+  'Inputs estilizados': {
+    practice: 'Estilize input, label, placeholder e estado de foco em um campo de email.',
+    exercise: 'Conclua quando o campo for legível, clicável e tiver foco visível.'
+  },
+  'Formulários bonitos': {
+    practice: 'Monte um formulário curto com duas colunas no desktop e uma coluna no mobile.',
+    exercise: 'Conclua quando labels, campos e botão formarem um fluxo claro de preenchimento.'
+  },
+  'Hover e estados visuais': {
+    practice: 'Adicione hover, focus e disabled em botões e links sem remover acessibilidade.',
+    exercise: 'Conclua quando mouse e teclado receberem feedback visual equivalente.'
+  },
+  'Introdução à responsividade': {
+    practice: 'Teste uma página em três larguras e liste os pontos onde o layout quebra.',
+    exercise: 'Conclua quando identificar ajustes reais antes de escrever media queries.'
+  },
+  'Media queries': {
+    practice: 'Adicione uma media query para mudar uma grade de duas colunas para uma coluna.',
+    exercise: 'Conclua quando a mudança acontecer no ponto em que o conteúdo começa a apertar.'
+  },
+  'Layout mobile-first': {
+    practice: 'Construa primeiro a versão mobile de uma seção e depois expanda para desktop.',
+    exercise: 'Conclua quando o CSS base funcionar no celular sem precisar desfazer estilos grandes.'
+  },
+  'Responsividade em textos': {
+    practice: 'Ajuste títulos, parágrafos e largura de leitura para celular e desktop.',
+    exercise: 'Conclua quando o texto não ficar largo demais no desktop nem grande demais no mobile.'
+  },
+  'Responsividade em imagens': {
+    practice: 'Faça imagens de cards se adaptarem à largura disponível sem distorcer.',
+    exercise: 'Conclua quando nenhuma imagem ultrapassar o container em telas pequenas.'
+  },
+  'Responsividade em cards': {
+    practice: 'Transforme uma grade de cards em uma coluna no mobile e múltiplas colunas no desktop.',
+    exercise: 'Conclua quando os cards continuarem com boa leitura em todas as larguras testadas.'
+  },
+  'Navbar responsiva': {
+    practice: 'Adapte uma navbar para empilhar links ou mostrar menu compacto em telas menores.',
+    exercise: 'Conclua quando todos os links continuarem acessíveis e com área de toque confortável.'
+  },
+  'Breakpoints comuns': {
+    practice: 'Teste 480px, 768px, 1024px e 1280px e defina ajustes baseados no conteúdo.',
+    exercise: 'Conclua quando cada breakpoint tiver motivo visual claro.'
+  },
+  'Ajustes para tablet': {
+    practice: 'Revise uma página em largura de tablet e ajuste colunas, espaçamentos e navegação.',
+    exercise: 'Conclua quando a tela média não parecer nem mobile esticado nem desktop espremido.'
+  },
+  'Projeto final com CSS responsivo': {
+    practice: 'Construa uma página completa com hero, cards, formulário e navbar responsiva.',
+    exercise: 'Conclua quando publicar a página e validar visualmente mobile, tablet e desktop.'
+  }
+}
