+++ Front Lab Academy/src/data/trilhas/frameworks-front-end/modulos.js
@@ -1,1 +1,11 @@
-export const frameworksFrontEndModules = []
+import { mkStarter } from '../../helpers/criar-starter.js'
+
+export const frameworksFrontEndModules = [
+      { title: 'Arquitetura por componentes', learn: ['composição', 'props', 'estado local'], practice: 'Separe uma tela de dashboard em Header, SummaryCard, FilterBar e DataList.', exercise: 'Conclua quando cada componente receber dados por props e tiver uma responsabilidade clara.', starter: mkStarter('Componentes', '#7b5cff', 'Componentes') },
+      { title: 'Roteamento em SPA', learn: ['rotas', 'parâmetros', 'layouts'], practice: 'Crie rotas para listagem, detalhe e página de erro em uma aplicação simples.', exercise: 'Conclua quando a rota de detalhe ler um parâmetro e renderizar o item correto.', starter: mkStarter('Roteamento', '#8b5cf6', 'Rotas') },
+      { title: 'Estado compartilhado', learn: ['lifting state', 'stores', 'contexto'], practice: 'Compartilhe o estado de carrinho entre lista de produtos, resumo e botão de limpar.', exercise: 'Conclua quando duas áreas da tela reagirem à mesma alteração de estado.', starter: mkStarter('Estado global', '#a78bfa', 'Estado') },
+      { title: 'Formulários e validação', learn: ['controlled/uncontrolled', 'regras de validação', 'feedback de erro'], practice: 'Construa um formulário de cadastro com validação de nome, email e senha.', exercise: 'Conclua quando erros aparecerem perto dos campos e o envio só ocorrer com dados válidos.', starter: mkStarter('Formulários', '#7c3aed', 'Form') },
+      { title: 'Consumo de APIs', learn: ['fetch em componentes', 'cache', 'retries'], practice: 'Crie uma listagem com busca, paginação simples, loading e estado vazio.', exercise: 'Conclua quando falhas de API não quebrarem a tela e mostrarem mensagem útil.', starter: mkStarter('API em framework', '#6d28d9', 'API framework') },
+      { title: 'Renderização e performance', learn: ['SSR/SSG/CSR', 'lazy loading', 'memoização'], practice: 'Identifique uma lista lenta e aplique lazy loading ou memoização onde fizer sentido.', exercise: 'Conclua quando houver comparação antes/depois com a mudança aplicada.', starter: mkStarter('Performance', '#7b5cff', 'Perf framework') },
+      { title: 'Estrutura de app escalável', learn: ['pastas por domínio', 'camadas', 'boas práticas de manutenção'], practice: 'Organize uma aplicação por domínio, separando components, services, types e pages.', exercise: 'Conclua quando um novo recurso puder ser adicionado sem misturar arquivos de domínios diferentes.', starter: mkStarter('Arquitetura app', '#8b5cf6', 'Arquitetura') }
+    ]
