+++ Front Lab Academy/src/data/trilhas/html/index.js
@@ -1,5 +1,10 @@
 import { htmlModules } from './modulos.js'
 
 export const htmlTrack = {
-  modules: htmlModules
-}
+    name: 'HTML', level: 'iniciante', levelLabel: 'Iniciante',
+    tags: ['HTML', 'Semântica', 'Formulários', 'SEO', 'Acessibilidade'],
+    accent: '#F97316', iconUrl: 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/html5/html5-original.svg',
+    description: 'Aprenda HTML do zero com 50 módulos, avançando dos fundamentos até páginas completas, formulários, semântica, SEO, acessibilidade e projetos práticos.',
+    modules: htmlModules,
+    challenge: { title: 'Desafio final: Site institucional semântico', brief: 'Construa um site institucional multi-seções com HTML semântico, formulário funcional, mídia incorporada, SEO básico e foco em acessibilidade.', portfolio: 'Entregáveis: deploy, auditoria A11y, revisão final e documentação da estrutura semântica.' }
+  }
