+++ Front Lab Academy/src/data/trilhas/javascript-e-logica/index.js
@@ -1,5 +1,10 @@
 import { javascriptELogicaModules } from './modulos.js'
 
 export const javascriptELogicaTrack = {
-  modules: javascriptELogicaModules
-}
+    name: 'JavaScript e lógica', level: 'intermediario', levelLabel: 'Básico ao intermediário',
+    tags: ['JavaScript', 'Lógica', 'DOM', 'Eventos', 'API'],
+    accent: '#FACC15', iconUrl: 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/javascript/javascript-original.svg',
+    description: 'Treine variáveis, operadores, condicionais, loops, arrays, objetos, funções, strings, datas, DOM, eventos, Fetch API e LocalStorage.',
+    modules: javascriptELogicaModules,
+    challenge: { title: 'Desafio final: Organizador de rotina inteligente', brief: 'Crie um app de rotina com tarefas, filtros, persistência local e integração com API para dados auxiliares.', portfolio: 'Entregáveis: app publicado, README técnico e vídeo de demonstração do fluxo completo.' }
+  }
