+++ Front Lab Academy/src/data/trilhas/javascript-e-logica/modulos.js
@@ -1,1 +1,11 @@
-export const javascriptELogicaModules = []
+import { mkStarter } from '../../helpers/criar-starter.js'
+
+export const javascriptELogicaModules = [
+      { title: 'Sintaxe e controle de fluxo', learn: ['variáveis', 'operadores', 'if/switch'], practice: 'Crie um verificador de idade e plano que retorne mensagens diferentes para visitante, aluno e assinante.', exercise: 'Conclua quando cada regra tiver uma condição clara e os resultados forem testados no console.', starter: mkStarter('Fluxo JS', '#eab308', 'Fluxo') },
+      { title: 'Loops e funções', learn: ['for/while', 'funções puras', 'escopo'], practice: 'Crie uma função que receba uma lista de notas e gere média, maior nota e situação final.', exercise: 'Conclua quando a função puder ser reutilizada com pelo menos três listas diferentes.', starter: mkStarter('Loops e funções', '#facc15', 'Loops') },
+      { title: 'Arrays e métodos', learn: ['map/filter/reduce', 'find/some/every', 'imutabilidade'], practice: 'Filtre uma lista de produtos por categoria, aplique desconto com map e calcule total com reduce.', exercise: 'Conclua quando a lista original continuar intacta e os resultados aparecerem separados.', starter: mkStarter('Arrays', '#f59e0b', 'Arrays') },
+      { title: 'Objetos e estruturas', learn: ['objetos aninhados', 'desestruturação', 'spread/rest'], practice: 'Modele um pedido com cliente, itens e endereço, depois gere um resumo usando desestruturação.', exercise: 'Conclua quando atualizar um campo com spread sem alterar o objeto original.', starter: mkStarter('Objetos', '#eab308', 'Objetos') },
+      { title: 'DOM e eventos', learn: ['seleção de elementos', 'delegação', 'renderização dinâmica'], practice: 'Crie uma lista de tarefas com adicionar, marcar como feita e remover usando eventos.', exercise: 'Conclua quando a interface atualizar sem recarregar a página.', starter: mkStarter('DOM e eventos', '#facc15', 'DOM') },
+      { title: 'Fetch API', learn: ['async/await', 'tratamento de erro', 'estado de carregamento'], practice: 'Busque uma lista de posts, mostre carregando, renderize os títulos e trate falha de rede.', exercise: 'Conclua quando a tela tiver estados de sucesso, carregamento e erro.', starter: mkStarter('Fetch API', '#fde047', 'API') },
+      { title: 'LocalStorage', learn: ['persistência local', 'serialização JSON', 'sincronização UI'], practice: 'Salve uma preferência de tema e uma lista simples de tarefas no navegador.', exercise: 'Conclua quando os dados reaparecerem corretamente após atualizar a página.', starter: mkStarter('LocalStorage', '#facc15', 'LocalStorage') }
+    ]
