+++ Front Lab Academy/src/pages/desafios.html
@@ -117,7 +117,7 @@
           </article>
           <article class="content-card">
             <h3>Validação</h3>
-            <p>Checklist final + evidência de funcionamento.</p>
+            <p>Revisão final + evidência de funcionamento.</p>
           </article>
         </div>
       </section>
@@ -148,6 +148,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
