+++ Front Lab Academy/src/pages/documentacao.html
@@ -472,6 +472,6 @@
       </p>
     </footer>
 
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
