+++ Front Lab Academy/src/pages/exercicios.html
@@ -116,7 +116,7 @@
             <p>Resolva em etapas pequenas e testáveis.</p>
           </article>
           <article class="content-card">
-            <h3>3. Revisar com checklist</h3>
+            <h3>3. Revisar com critérios</h3>
             <p>Valide acessibilidade, responsividade e qualidade de código.</p>
           </article>
         </div>
@@ -148,6 +148,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
