+++ Front Lab Academy/src/pages/index.html
@@ -393,6 +393,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
