+++ Front Lab Academy/src/pages/mini-ide.html
@@ -107,11 +107,11 @@
             <h3>Fluxo recomendado</h3>
             <p>
               1) Leia o objetivo do módulo. 2) Implemente mínimo viável. 3)
-              Rode. 4) Refatore. 5) Compare com checklist.
+              Rode. 4) Refatore. 5) Compare com os critérios.
             </p>
           </article>
           <article class="content-card">
-            <h3>Checklist de qualidade</h3>
+            <h3>Critérios de qualidade</h3>
             <p>
               Semântica, responsividade, estados de erro, legibilidade de código
               e separação de responsabilidades.
@@ -189,6 +189,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
