+++ Front Lab Academy/src/pages/modulos.html
@@ -160,6 +160,6 @@
       </p>
     </footer>
 
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
