+++ Front Lab Academy/src/pages/pratica.html
@@ -122,6 +122,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
