+++ Front Lab Academy/src/pages/progresso.html
@@ -137,6 +137,6 @@
       </p>
     </footer>
 
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
