+++ Front Lab Academy/src/pages/projetos.html
@@ -156,6 +156,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
