+++ Front Lab Academy/src/pages/roteiro.html
@@ -135,6 +135,6 @@
       </p>
     </footer>
 
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
