+++ Front Lab Academy/src/pages/script.js
@@ -1,0 +1,1 @@
+import '../script.js'
