+++ Front Lab Academy/src/pages/sobre.html
@@ -149,7 +149,7 @@
             <h3>Portfólio com evidência</h3>
             <p>
               Ao concluir uma trilha, a pessoa não leva só anotações: leva
-              projetos, checklists, esqueleto de solução, evolução registrada e
+              projetos, critérios de revisão, esqueleto de solução, evolução registrada e
               material que pode ser apresentado em entrevistas, freelas e
               oportunidades reais.
             </p>
@@ -203,6 +203,6 @@
         >
       </p>
     </footer>
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
