+++ Front Lab Academy/src/pages/trilhas.html
@@ -159,7 +159,6 @@
       </p>
     </footer>
 
-    <script type="module" src="../script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
-
