+++ Front Lab Academy/src/styles/style.css
@@ -1491,12 +1491,25 @@ pre {
   padding: 0.9rem;
 }
 
+.module-study-section {
+  display: grid;
+  gap: 0.55rem;
+  border: 1px solid color-mix(in srgb, var(--border) 74%, transparent);
+  border-radius: 0.7rem;
+  background: color-mix(in srgb, var(--card) 74%, var(--surface-soft));
+  padding: 0.85rem;
+}
+
 .module-study-content h5 {
   margin: 0.3rem 0 0;
   color: color-mix(in srgb, var(--track-accent, var(--blue)) 58%, var(--text-main));
   font-size: 0.82rem;
   font-weight: 900;
   text-transform: uppercase;
+}
+
+.module-study-section h5 {
+  margin: 0;
 }
 
 .module-study-content p {
@@ -1506,31 +1519,70 @@ pre {
   line-height: 1.7;
 }
 
+.module-study-list {
+  display: grid;
+  gap: 0.45rem;
+  margin: 0;
+  padding-left: 1.15rem;
+  color: var(--text-secondary);
+  font-weight: 500;
+  line-height: 1.6;
+}
+
+.module-study-raw {
+  display: grid;
+  gap: 0.75rem;
+}
+
+.module-study-demo {
+  display: grid;
+  gap: 0.65rem;
+}
+
+.module-study-preview {
+  min-height: 240px;
+  border-radius: 0.65rem;
+}
+
 .module-study-content pre {
   margin: 0;
   overflow: auto;
   border: 1px solid var(--border);
   border-radius: 0.65rem;
-  background: var(--code-bg);
+  background: #0D1117;
   padding: 0.85rem;
 }
 
-.module-study-content code {
-  color: var(--code-text);
+.module-study-content pre code {
+  color: #D6DEFA;
   font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
   font-size: 0.8rem;
   line-height: 1.55;
   white-space: pre;
 }
 
-.module-progress-checklist {
+.module-study-content pre .code-tag,
+.module-study-content pre .code-attr,
+.module-study-content pre .code-string,
+.module-study-content pre .code-comment,
+.module-study-content pre .code-number,
+.module-study-content pre .code-keyword {
+  color: inherit;
+}
+
+.module-progress-control {
   display: grid;
-  gap: 0.85rem;
+  gap: 0.8rem;
   margin-top: 1rem;
-  border: 1px solid color-mix(in srgb, var(--track-accent, var(--border)) 36%, var(--border));
-  border-radius: 0.85rem;
-  background: color-mix(in srgb, var(--track-accent, var(--surface-soft)) 7%, var(--surface-soft));
-  padding: 1rem;
+  border: 1px solid var(--border);
+  border-radius: 0.75rem;
+  background: color-mix(in srgb, var(--surface-soft) 82%, var(--card));
+  padding: 0.85rem;
+}
+
+.module-progress-control:has(.is-complete) {
+  border-color: color-mix(in srgb, var(--green) 42%, var(--border));
+  background: color-mix(in srgb, var(--green) 8%, var(--surface-soft));
 }
 
 .module-progress-head {
@@ -1547,13 +1599,13 @@ pre {
 
 .module-progress-head h4 {
   color: var(--text-main);
-  font-size: 1rem;
+  font-size: 0.92rem;
 }
 
 .module-progress-head p {
   margin-top: 0.2rem;
   color: var(--text-muted);
-  font-size: 0.84rem;
+  font-size: 0.8rem;
 }
 
 .module-progress-head > span {
@@ -1570,6 +1622,18 @@ pre {
   border-color: color-mix(in srgb, var(--green) 50%, var(--border));
   background: color-mix(in srgb, var(--green) 14%, var(--surface-soft));
   color: var(--green);
+}
+
+.toggle-module-completion {
+  justify-self: end;
+  min-height: 34px;
+  padding: 0.45rem 0.8rem;
+  background: transparent;
+}
+
+.module-progress-control:has(.is-complete) .toggle-module-completion {
+  border-color: color-mix(in srgb, var(--green) 44%, var(--border));
+  color: color-mix(in srgb, var(--green) 70%, var(--text-main));
 }
 
 .module-ide {
@@ -1633,7 +1697,7 @@ pre {
 
 .module-ide-grid textarea[data-type="html"],
 #editor[data-active-tab="html"] {
-  color: #9A3412;
+  color: #334155;
 }
 
 .module-ide-grid textarea[data-type="css"],
@@ -1643,7 +1707,7 @@ pre {
 
 .module-ide-grid textarea[data-type="js"],
 #editor[data-active-tab="js"] {
-  color: #854D0E;
+  color: #475569;
 }
 
 :root[data-theme="dark"] .module-ide-grid textarea {
@@ -1862,6 +1926,16 @@ pre {
 :root[data-theme="dark"] .module-learning-grid {
   border-color: rgba(29, 41, 77, 0.72);
   background: color-mix(in srgb, var(--card) 86%, var(--purple) 4%);
+}
+
+:root[data-theme="dark"] .module-progress-control {
+  border-color: rgba(49, 65, 104, 0.75);
+  background: color-mix(in srgb, var(--card) 92%, var(--bg-main));
+}
+
+:root[data-theme="dark"] .module-progress-control:has(.is-complete) {
+  border-color: color-mix(in srgb, var(--green) 46%, var(--border));
+  background: color-mix(in srgb, var(--green) 10%, var(--card));
 }
 
 :root[data-theme="dark"] .module-ide,
